+++ prototype.creep.js
@@ -15,6 +15,21 @@ Creep.prototype.logic = function() {
   roles[this.memory.role].run(this);
 }
 
+// Returns true if the Creep is ready for work, false otherwise
+Creep.prototype.isCharged = function() {
+
+  // Change internal status when appropriate
+  if (this.carry.energy <= 0) {
+    this.memory.charged = false;
+
+  } else if (this.carry.energy == this.carryCapacity) {
+    this.memory.charged = true;
+  }
+
+  // Return true/false accordingly
+  return this.memory.charged;
+}
+
 // Boolean Boolean -> void
 Creep.prototype.recharge = function(useSource, useStorage) {
 
@@ -50,5 +65,81 @@ Creep.prototype.recharge = function(useSource, useStorage) {
       // Otherwise get closer
       this.moveTo(source);
     }
+  }
+}
+
+// Boolean -> void
+Creep.prototype.longRecharge = function(pickUpDroppedResources) {
+
+  if (pickUpDroppedResources) {
+
+    // Scan for any dropped resources nearby
+    const dropped_resources = this.pos.findInRange(FIND_DROPPED_RESOURCES,10);
+
+    // If found any
+    if (dropped_resources.length) {
+
+      // Find closest one
+      let dropped_resource = this.pos.findClosestByRange(FIND_DROPPED_RESOURCES);
+
+      // Go get it!
+      if (this.pickup(dropped_resource) == ERR_NOT_IN_RANGE) {
+        this.moveTo(dropped_resource);
+
+        // Do not go for harvesting energy sources
+        return;
+      }
+    }
+  }
+
+  // If Creep is in target room
+  if (this.room.name == this.memory.target) {
+
+    // Find active source
+    const source = this.pos.findClosestByPath(FIND_SOURCES_ACTIVE);
+
+    // Try harvesting from source
+    if (this.harvest(source) == ERR_NOT_IN_RANGE) {
+      // Otherwise get closer to source
+      this.moveTo(source, {
+        visualizePathStyle: { stroke: '#7fffd4' } // aquamarine
+      });
+    }
+
+  // If not in target room, find exit and move towards it
+  } else {
+    const exit = this.room.findExitTo(this.memory.target);
+    this.moveTo(this.pos.findClosestByPath(exit));
+  }
+}
+
+Creep.prototype.transferEnergyToStructure = function() {
+
+  let structure = this.pos.findClosestByPath(FIND_MY_STRUCTURES, {
+    // Filter results by appropriate structure type, as follow
+    filter: structure => ((
+      structure.structureType == STRUCTURE_SPAWN
+      || structure.structureType == STRUCTURE_EXTENSION
+      || structure.structureType == STRUCTURE_TOWER)
+      && structure.energy < structure.energyCapacity
+    )
+  });
+
+  // Backup to storage if all other structure are fully charged
+  if (!structure) structure = this.room.storage;
+
+  // If structure found, go transfer energy to it
+  if (structure) {
+
+    // try transfer energy to it
+    if (this.transfer(structure, RESOURCE_ENERGY) == ERR_NOT_IN_RANGE) {
+
+      // if not in range, move closer
+      this.moveTo(structure);
+    }
+
+  // If any target available, try to upgrade controller
+  } else {
+    upgrader.run(creep);
   }
 }